Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

ADNLPModelProblems.meta

ADNLPModelProblems.generate_meta

​

ADNLPModelProblems.meta — Constant

ADNLPModelProblems.meta

A composite type that represents the main features of the optimization problem. optimize obj(x) subject to lvar ≤ x ≤ uvar lcon ≤ cons(x) ≤ ucon –- The following keys are valid: Problem meta

nvar: number of general constraints

variable_size: true if we can modify problem size

ncon: number of general constraints

variable_con_size: true if we can modify problem size

nnzo: number of nonzeros in all objectives gradients

nnzh: number of elements needed to store the nonzeros in the sparse Hessian

nnzj: number of elements needed to store the nonzeros in the sparse Jacobian

minimize: true if optimize == minimize

name: problem name

Solution meta

optimalvalue: best known objective value (NaN if unknown, -Inf if unbounded problem)

has_multiple_solution: true if the problem has more than one global solution

is_infeasible: true if problem is infeasible

Classification

objtype: #in objtypes

contype: #in contypes

origin: #in origins

deriv: UInt8 with the largest order derivatives available

not_everywhere_defined: true if the problem might return NaN or Inf outside the bounds

has_cvx_obj: true if the problem has a convex objective

has_cvx_con: true if the problem has convex constraints

has_equalities_only: true if the problem constraints are equality constraints (doesn't include bounds)

has_inequalities_only: true if the problem constraints are inequality constraints (doesn't include bounds)

has_bounds: true if the problem has bound constraints

has_fixed_variables: true if it has fixed variables

cqs: Between 0 and 4 indicates the constraint qualification of the problem, see cqs(i) for the correspondance.

ADNLPModelProblems.generate_meta — Method

generate_meta(jmodel, name, variable_size, variable_con_size, cvx_obj, cvx_con, quad_cons) generate_meta(name, variable_size, variable_con_size, cvx_obj, cvx_con, quad_cons) is used to generate the meta of a given JuMP model.